bug fixes in dlsch functions

+++ targets/SIMU/USER/init_lte.h
@@ -24,7 +24,7 @@
 
 PHY_VARS_eNB* init_lte_eNB(LTE_DL_FRAME_PARMS *frame_parms,
                            uint8_t eNB_id,
-                           uint8_t Nid_cell,
+                           uint16_t Nid_cell,
 			   eNB_func_t node_function,
                            uint8_t abstraction_flag);
